requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-anthropic>=0.1.0
langchain-groq>=0.0.1
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
